# fix: check for non-empty accept header before performing accept header flow