Closes #635

Before:

```
Running 5s test @ http://localhost:1447?test=test
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   162.52ms  253.89ms   1.99s    92.58%
    Req/Sec   473.68     72.50   643.00     73.00%
  4796 requests in 5.09s, 2.54MB read
  Non-2xx or 3xx responses: 4796
Requests/sec:    941.74
Transfer/sec:    511.34KB
```

After:

```
Running 5s test @ http://localhost:1447?test=test
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.74ms    5.18ms 124.77ms   99.08%
    Req/Sec    14.80k     1.48k   15.78k    90.20%
  150207 requests in 5.10s, 19.63MB read
Requests/sec:  29431.85
Transfer/sec:      3.85MB
```